+++ libs/dot-components/src/lib/components/link/Link.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Story, Meta } from '@storybook/react/types-6-0';
 
 import { DotLink, LinkProps } from './Link';
-import { Typography } from '@material-ui/core';
+import { DotTypography } from '../typography/Typography';
 
 export default {
   title: 'Components/Link',
@@ -14,8 +14,10 @@ export default {
 
 export const Default: Story<LinkProps> = (args) => <DotLink {...args} />;
 export const LinkInParagraph: Story<LinkProps> = (args) => (
-  <Typography variant="body1">
-    This is a <DotLink {...args} /> that will take you to a page that has
-    nothing to do with Batman!
-  </Typography>
+  <DotTypography variant="body1">
+    <>
+      This is a <DotLink {...args} /> that will take you to a page that has
+      nothing to do with Batman!
+    </>
+  </DotTypography>
 );