+++ libs/dot-components/src/lib/components/typography/Typography.spec.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { renderWithTheme as render } from '../../testing-utils/RenderWithTheme';
+import { DotTypography, TypographyProps } from './Typography';
+
+describe('DotTypography', () => {
+  it('should have unchanged API', () => {
+    const props = {
+      children: 'some text',
+      component: 'h1',
+      variant: 'h1',
+    };
+    const typographyProps: TypographyProps = {
+      children: 'some text',
+      component: 'h1',
+      variant: 'h1',
+    };
+    expect(typographyProps).toEqual(props);
+  });
+
+  it('should render successfully', () => {
+    const { baseElement } = render(<DotTypography children="Testing" />);
+    expect(baseElement).toBeTruthy();
+  });
+});